Eliminate va_list using functions.
This type is opaque and handled by C macros, which are unavailable anyway. And this will break on
some systems when trying to treat in as an opaque value, because apparently on some systems it is a
structure passed by value larger than a dword.

In any case it is always possible to use a foreign-funcall to call the variadic versions of the
functions directly, if for some reason multiple calls to single-argument functions are wrong.